selfdemo-admin-api/src/test/java/com/saul/springboot/selfDemo/applications/ReviewServiceTests.java
package com.saul.springboot.selfDemo.applications;


import com.saul.springboot.selfDemo.domain.Review;
import com.saul.springboot.selfDemo.domain.ReviewRepository;
import org.junit.Before;
import org.junit.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import java.util.Arrays;
import java.util.List;

import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;
import static org.mockito.BDDMockito.given;

public class ReviewServiceTests {

    private ReviewService reviewService;

    @Mock
    ReviewRepository reviewRepository;

    @Before
    public void setUp() {
        MockitoAnnotations.initMocks(this);

        reviewService = new ReviewService(reviewRepository);
    }

    @Test
    public void getReviews() {
        given(reviewRepository.findAll())
                .willReturn(Arrays.asList(Review.builder().writer("poppo").build()));

        List<Review> reviews = reviewService.getReviews();
        Review review = reviews.get(0);

        assertThat(review.getWriter(), is("poppo"));
    }

}
